refactor: split thread command into focused modules
