+++ src/commands/thread.ts
@@ -1,348 +1,8 @@
-import type { TwistApi } from '@doist/twist-sdk'
-import chalk from 'chalk'
 import { Command, Option } from 'commander'
-import { getTwistClient } from '../lib/api.js'
 import { withUnvalidatedChoices } from '../lib/completion.js'
-import { formatRelativeDate } from '../lib/dates.js'
-import { openEditor, readStdin } from '../lib/input.js'
-import { renderMarkdown } from '../lib/markdown.js'
-import type { MutationOptions, PaginatedViewOptions } from '../lib/options.js'
-import { colors, formatJson, isAccessible } from '../lib/output.js'
-import { assertChannelIsPublic } from '../lib/public-channels.js'
-import { extractId, parseRef, resolveThreadId } from '../lib/refs.js'
-
-type ViewOptions = PaginatedViewOptions & { comment?: string; unread?: boolean; context?: string }
-
-type ReplyOptions = MutationOptions & { notify?: string }
-
-type DoneOptions = MutationOptions
-
-function printSeparator(label: string): void {
-    const totalWidth = 60
-    const labelWithPadding = ` ${label} `
-    const remainingWidth = totalWidth - labelWithPadding.length
-    const leftWidth = Math.floor(remainingWidth / 2)
-    const rightWidth = remainingWidth - leftWidth
-    const dashChar = isAccessible() ? '-' : '─'
-    const line = chalk.dim(
-        dashChar.repeat(leftWidth) + labelWithPadding + dashChar.repeat(rightWidth),
-    )
-    console.log('')
-    console.log(line)
-    console.log('')
-}
-
-function pluralize(count: number, singular: string): string {
-    return count === 1 ? singular : `${singular}s`
-}
-
-interface CommentLike {
-    id: number
-    creator: number
-    posted: Date
-    content: string
-}
-
-function printComment(comment: CommentLike, userMap: Map<number, string>, raw: boolean): void {
-    const author = colors.author(userMap.get(comment.creator) || `user:${comment.creator}`)
-    const time = colors.timestamp(formatRelativeDate(comment.posted))
-    console.log(`${author}  ${time}  ${colors.timestamp(`id:${comment.id}`)}`)
-    console.log(raw ? comment.content : renderMarkdown(comment.content))
-    console.log('')
-}
-
-async function viewSingleComment(
-    client: TwistApi,
-    threadId: number,
-    commentId: number,
-    options: ViewOptions,
-): Promise<void> {
-    const [threadResponse, commentResponse] = await client.batch(
-        client.threads.getThread(threadId, { batch: true }),
-        client.comments.getComment(commentId, { batch: true }),
-    )
-
-    const thread = threadResponse.data
-    const comment = commentResponse.data
-
-    const userIds = new Set([thread.creator, comment.creator])
-    const userCalls = [...userIds].map((id) =>
-        client.workspaceUsers.getUserById(
-            { workspaceId: thread.workspaceId, userId: id },
-            { batch: true },
-        ),
-    )
-    const [channelResponse, ...userResponses] = await client.batch(
-        client.channels.getChannel(thread.channelId, { batch: true }),
-        ...userCalls,
-    )
-
-    const channel = channelResponse.data
-    const userMap = new Map(userResponses.map((r) => [r.data.id, r.data.name]))
-
-    if (options.json) {
-        const output = {
-            ...comment,
-            creatorName: userMap.get(comment.creator),
-            channelName: channel.name,
-            threadTitle: thread.title,
-        }
-        console.log(formatJson(output, undefined, options.full))
-        return
-    }
-
-    if (options.ndjson) {
-        console.log(
-            JSON.stringify({
-                type: 'comment',
-                ...comment,
-                creatorName: userMap.get(comment.creator),
-            }),
-        )
-        return
-    }
-
-    console.log(chalk.bold(thread.title))
-    console.log(colors.channel(`[${channel.name}]`))
-    console.log('')
-    printComment(comment, userMap, options.raw ?? false)
-}
-
-async function viewThread(ref: string, options: ViewOptions): Promise<void> {
-    const parsed = parseRef(ref)
-    const threadId = resolveThreadId(ref)
-    const urlCommentId = parsed.type === 'url' ? parsed.parsed.commentId : undefined
-    let commentId: number | undefined
-    if (options.comment !== undefined) {
-        commentId = extractId(options.comment)
-    } else {
-        commentId = urlCommentId
-    }
-    const client = await getTwistClient()
-
-    if (commentId !== undefined) {
-        return viewSingleComment(client, threadId, commentId, options)
-    }
-
-    const limit = options.limit ? parseInt(options.limit, 10) : 50
-
-    const [threadResponse, commentsResponse] = await client.batch(
-        client.threads.getThread(threadId, { batch: true }),
-        client.comments.getComments(
-            {
-                threadId,
-                from: options.since ? new Date(options.since) : undefined,
-                limit,
-            },
-            { batch: true },
-        ),
-    )
-
-    const thread = threadResponse.data
-    const comments = commentsResponse.data
-
-    await assertChannelIsPublic(thread.channelId, thread.workspaceId)
-
-    let lastReadObjIndex: number | null = null
-    if (options.unread) {
-        const unreadData = await client.threads.getUnread(thread.workspaceId)
-        const threadUnread = unreadData.find((u) => u.threadId === threadId)
-        if (!threadUnread) {
-            console.log('No unread comments in this thread.')
-            return
-        }
-        lastReadObjIndex = threadUnread.objIndex
-    }
-
-    const userIds = new Set<number>([thread.creator, ...comments.map((c) => c.creator)])
-    const userCalls = [...userIds].map((id) =>
-        client.workspaceUsers.getUserById(
-            { workspaceId: thread.workspaceId, userId: id },
-            { batch: true },
-        ),
-    )
-    const [channelResponse, ...userResponses] = await client.batch(
-        client.channels.getChannel(thread.channelId, { batch: true }),
-        ...userCalls,
-    )
-
-    const channel = channelResponse.data
-    const userMap = new Map(userResponses.map((r) => [r.data.id, r.data.name]))
-
-    if (options.json) {
-        const output = {
-            thread: {
-                ...thread,
-                channelName: channel.name,
-                creatorName: userMap.get(thread.creator),
-            },
-            comments: comments.map((c) => ({
-                ...c,
-                creatorName: userMap.get(c.creator),
-            })),
-        }
-        console.log(formatJson(output, undefined, options.full))
-        return
-    }
-
-    if (options.ndjson) {
-        const threadOutput = {
-            type: 'thread',
-            ...thread,
-            channelName: channel.name,
-            creatorName: userMap.get(thread.creator),
-        }
-        console.log(JSON.stringify(threadOutput))
-        for (const c of comments) {
-            console.log(
-                JSON.stringify({ type: 'comment', ...c, creatorName: userMap.get(c.creator) }),
-            )
-        }
-        return
-    }
-
-    console.log(chalk.bold(thread.title))
-    console.log(colors.channel(`[${channel.name}]`))
-    console.log('')
-
-    if (options.unread && lastReadObjIndex !== null) {
-        const contextSize = options.context ? parseInt(options.context, 10) : 0
-        const unreadComments = comments.filter((c) => (c.objIndex ?? 0) > lastReadObjIndex)
-        const contextComments = comments
-            .filter((c) => (c.objIndex ?? 0) <= lastReadObjIndex)
-            .sort((a, b) => (b.objIndex ?? 0) - (a.objIndex ?? 0))
-            .slice(0, contextSize)
-            .reverse()
-
-        if (unreadComments.length === 0) {
-            console.log('No unread comments.')
-            return
-        }
-
-        const creatorName = userMap.get(thread.creator) || `user:${thread.creator}`
-        console.log(
-            `${colors.author(creatorName)}  ${colors.timestamp(formatRelativeDate(thread.posted))}  ${chalk.dim('(original post)')}`,
-        )
-        console.log('')
-        console.log(options.raw ? thread.content : renderMarkdown(thread.content))
-
-        if (contextComments.length > 0) {
-            const firstContextIndex = contextComments[0].objIndex ?? 0
-            const skippedCount = firstContextIndex - 1
-            if (skippedCount > 0) {
-                printSeparator(`${skippedCount} ${pluralize(skippedCount, 'comment')} skipped`)
-            } else {
-                console.log('')
-            }
-            for (const comment of contextComments) {
-                printComment(comment, userMap, options.raw ?? false)
-            }
-        } else if (lastReadObjIndex > 0) {
-            printSeparator(`${lastReadObjIndex} ${pluralize(lastReadObjIndex, 'comment')} skipped`)
-        }
-
-        printSeparator(`UNREAD (${unreadComments.length} new)`)
-
-        for (const comment of unreadComments) {
-            printComment(comment, userMap, options.raw ?? false)
-        }
-    } else {
-        const creatorName = userMap.get(thread.creator) || `user:${thread.creator}`
-        console.log(
-            `${colors.author(creatorName)}  ${colors.timestamp(formatRelativeDate(thread.posted))}`,
-        )
-        console.log('')
-        console.log(options.raw ? thread.content : renderMarkdown(thread.content))
-        console.log('')
-
-        if (comments.length > 0) {
-            console.log(
-                chalk.dim(`--- ${comments.length} ${pluralize(comments.length, 'comment')} ---`),
-            )
-            console.log('')
-
-            for (const comment of comments) {
-                printComment(comment, userMap, options.raw ?? false)
-            }
-        }
-    }
-}
-
-async function replyToThread(
-    ref: string,
-    content: string | undefined,
-    options: ReplyOptions,
-): Promise<void> {
-    const threadId = resolveThreadId(ref)
-
-    let replyContent = await readStdin()
-    if (!replyContent && content) {
-        replyContent = content
-    }
-    if (!replyContent) {
-        replyContent = await openEditor()
-    }
-    if (!replyContent || replyContent.trim() === '') {
-        console.error('No content provided.')
-        process.exit(1)
-    }
-
-    const notifyValue = options.notify ?? 'EVERYONE_IN_THREAD'
-    let recipients: string | number[]
-    if (notifyValue === 'EVERYONE' || notifyValue === 'EVERYONE_IN_THREAD') {
-        recipients = notifyValue
-    } else {
-        recipients = notifyValue.split(',').map((userRef) => {
-            const trimmed = userRef.trim()
-            if (!trimmed) {
-                console.error('Invalid user reference list: found empty value')
-                process.exit(1)
-                return 0
-            }
-            try {
-                return extractId(trimmed)
-            } catch {
-                console.error(`Invalid user reference: ${trimmed}. Use 123 or id:123`)
-                process.exit(1)
-                return 0
-            }
-        })
-    }
-
-    if (options.dryRun) {
-        console.log('Dry run: would post comment to thread', threadId)
-        console.log(`Notify: ${Array.isArray(recipients) ? recipients.join(', ') : recipients}`)
-        console.log('')
-        console.log(replyContent)
-        return
-    }
-
-    const client = await getTwistClient()
-    const thread = await client.threads.getThread(threadId)
-    await assertChannelIsPublic(thread.channelId, thread.workspaceId)
-    const comment = await client.comments.createComment({
-        threadId,
-        content: replyContent,
-        recipients,
-    } as Parameters<typeof client.comments.createComment>[0])
-
-    console.log(`Comment posted: ${comment.url}`)
-}
-
-async function markThreadDone(ref: string, options: DoneOptions): Promise<void> {
-    const threadId = resolveThreadId(ref)
-
-    if (options.dryRun) {
-        console.log(`Dry run: would archive thread ${threadId}`)
-        return
-    }
-
-    const client = await getTwistClient()
-    const thread = await client.threads.getThread(threadId)
-    await assertChannelIsPublic(thread.channelId, thread.workspaceId)
-    await client.inbox.archiveThread(threadId)
-    console.log(`Thread ${threadId} archived.`)
-}
+import { markThreadDone } from './thread/mutate.js'
+import { replyToThread } from './thread/reply.js'
+import { viewThread } from './thread/view.js'
 
 export function registerThreadCommand(program: Command): void {
     const thread = program.command('thread').description('Thread operations')
